requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2==2.9.9
mysqlclient==2.2.4
redis==5.0.2
requests==2.31.0
httpx==0.27.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-jaeger==1.21.0
#opentelemetry-exporter-prometheus==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-psycopg2==0.44b0
opentelemetry-instrumentation-requests==0.44b0
opentelemetry-instrumentation-redis==0.44b0
opentelemetry-instrumentation-mysqlclient==0.44b0
opentelemetry-exporter-otlp==1.23.0